fix (issues_v1): move Paging and SearchPostsFilter to dto package
