+++ app/src/main/java/ru/blog/services/ports/PostsRepository.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import org.springframework.data.domain.PageRequest;
 import ru.blog.models.PostDetails;
 import ru.blog.models.PostPreview;
-import ru.blog.models.SearchPostsFilter;
+import ru.blog.dto.SearchPostsFilter;
 
 public interface PostsRepository {
   String save(PostDetails postDetails);
